+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.0.28d - 21/09/2024
 
 - Added `.github/workflows` containing platform-specific workflows
+- Fixed `Worker` tests
 
 ## 0.0.27d - 12/09/2024